perf(rollup): update bundle management
Changes to be committed:
modified:   demo/index.html
modified:   index.js
modified:   package-lock.json
modified:   package.json
deleted:    rollup.config.js
new file:   rollup.config.mjs
modified:   src/auro-badge.js
modified:   test/auro-badge.test.js
modified:   .github/workflows/testPublish.yml
