+++ demo/index.html
@@ -25,7 +25,7 @@
           Prism.highlightAll();
         });
     </script>
-    <script type="module" src="../src/auro-badge.js"></script>
+    <script type="module" src="../index.js"></script>
     <script src="https://unpkg.com/@alaskaairux/auro-accordion@latest/dist/auro-accordion__bundled.js" type="module"></script>
   </body>
 </html>
