+++ index.js
@@ -1,1 +1,17 @@
 export * from './dist/auro-badge';
+
+import { AuroBadge } from './src/auro-badge.js';
+
+/**
+ * Register Custom Element.
+ * @param {Object} name - Name to use for custom element.
+ * @returns {void}
+ */
+export function registerComponent(name = 'custom-badge') {
+  // alias definition
+  if (!customElements.get(name)) {
+    customElements.define(name, class extends AuroBadge {});
+  }
+}
+
+registerComponent('auro-badge');
