+++ package.json
@@ -41,10 +41,7 @@
     "@commitlint/cli": "^11.0.0",
     "@commitlint/config-conventional": "^11.0.0",
     "@open-wc/testing": "^3.1.7",
-    "@rollup/plugin-alias": "^3.1.2",
-    "@rollup/plugin-babel": "^5.3.0",
-    "@rollup/plugin-commonjs": "^17.1.0",
-    "@rollup/plugin-node-resolve": "^11.2.0",
+    "@rollup/plugin-node-resolve": "^15.0.1",
     "@semantic-release/changelog": "^5.0.1",
     "@semantic-release/git": "^9.0.0",
     "@semantic-release/npm": "^7.0.10",
@@ -72,10 +69,8 @@
     "postcss-remove-rules": "^1.0.0",
     "postcss-selector-replace": "^1.0.2",
     "prismjs": "^1.23.0",
-    "rollup": "^2.40.0",
-    "rollup-plugin-minify-html-literals": "^1.2.5",
-    "rollup-plugin-serve": "^1.1.0",
-    "rollup-plugin-terser": "^7.0.2",
+    "rollup": "^3.19.1",
+    "rollup-plugin-serve": "^2.0.2",
     "semantic-release": "^17.4.0",
     "sinon": "^9.2.4",
     "stylelint": "^13.11.0",
